% Generated by roxygen2: do not edit by hand
\name{BuildWaddington}
\alias{BuildWaddington}
\title{Build Waddington Landscape}
\usage{
BuildWaddington(dcALL.m,epiCT.idx,ccat.v,theta,phi,selCells.idx,x.v,y.v)
}

\arguments{

\item{dcALL.m}{The diffusion map matrix with rows labeling cell and columns labeling DCs.}

\item{epiCT.idx}{An index vector specifying the cellular states, which should be of the same length as rows of `dcALL.m`.}

\item{ccat.v}{The cellular plasticity estimates for the same cells as
  specified in the previous arguments}

\item{theta}{The longitude angle to specify viewpoint of landscape}

\item{phi}{The polar angle to specify viewpoint of landscape}

\item{selCells.idx}{An index vector specifying the cells to be shown}

\item{x.v}{The grid x (DC1) positions where to evaluate the landscape}

\item{y.v}{The grid y (DC2) positions where to evaluate the landscape}          

}


\value{

}
\description{
This function takes the estimated diffusion map and cellular plasticity
values, and builds a corresponding Waddington landscape. Of note, this
function may need to be tweaked to alter the region to be plotted,
if particular features of the landscape want to be emphasized.
}
\examples{

}
\references{

}
