% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompTFIL.R
\name{CompTFIL}
\alias{CompTFIL}
\title{Computation of TFIL}
\usage{
CompTFIL(tfa.m, pheno.v, thDTFA = "BF", thHits = 0.05)
}
\arguments{
\item{tfa.m}{The TFA-matrix defined over tissue-specific TFs (rows) and cells (columns)}

\item{pheno.v}{A phenotype integer vector for the cells, with normal cells indexed as 0, and
preneoplastic cells indexed with 1. Cells from more advanced stages including
cancer are indexed with positive integers larger than 1, with higher the integer
the more advanced the disease stage is. Note that this vector must have at least
2 disease stages (0,1) representing normal and preneoplastic cells.}

\item{thDTFA}{The significance threshold for defining differentially active TFs
between normal and preneoplastic cells. #' By default this is
Bonferroni-adjusted P<0.05, but can be numericm#' in which case the
provided value is used.}

\item{thHits}{The significance threshold on P-values for declaring inactivated TFs
for each cell. By default this value is 0.05.}
}
\value{
A list with the following elements

tfil
The transcription factor inactivation load for all cells. In practice we are
interested in those of the preneoplastic cells only, but for
convenience we return the load for all cells.

tfif
As the TFIL, but expressed as a fraction of the total number
of tissue-specific TFs

iTF
A vector specifying the gene-identifiers of the TFs that
are inactivated.

hitMatrix
A binary 0-1 matrix defined over the inactivated TFs and all cells, with 1
indicating that the given TF is significantly inactivated in the given cell.

tfa
The TFA-matrix input object.
}
\description{
This function takes as input the estimated TFA matrix and computes for each
cell the transcription factor inactivation load (TFIL), defined as the number
of tissue-specific TFs that are inactivated in a given cell compared to the
normal cells.
}
\examples{

 
}
\references{

}
