% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTFA.R
\name{EstTFA}
\alias{EstTFA}
\title{Estimation of Transcription Factor Regulatory Activity (TFA)}
\usage{
EstTFA(data, reg.m, norm = c("z", "c"), ncores = 4)
}
\arguments{
\item{data}{A scRNA-Seq data matrix with rows labeling genes and columns 
labeling single cells and with rownames annotated to a gene-identifier.
scRNA-Seq data matrix should have undergone prior QC to remove poor quality
cells and each cell normalized by library size. If data has not been log-transformed,
the function will log2-transform with a pseudocount of +1.}

\item{reg.m}{The regulon matrix with rows labeling gene-targets (same gene-identifier as
in \code{exp.m}) and columns labeling TFs. Each column has entries +1, -1 or 0,
depending on whether the gene is positively, negatively or not regulated by the
given TF.}

\item{norm}{Type of normalization to use if data is provided as a matrix.
"z" stands for z-score transformation, "c" just row centres the data matrix.}

\item{ncores}{Number of parallel cores to use}
}
\value{
tfa.m
The estimated TFA-matrix with rows labeling TFs and columns labeling cells
}
\description{
This function takes as input a log-normalized scRNA-Seq data matrix and a
set of TF-regulons and estimates for each TF the corresponding TF
regulatory activity level (TFA) in each cell.
}
\examples{

}
\references{
Teschendorff AE., Wang N. \emph{Improved detection of tumor suppressor
  events in single cell RNA-Seq data}. NPJ Genom Med. 2020 Oct 7;5:43.
}
