% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDMAP.R
\name{PlotDMAP}
\alias{PlotDMAP}
\title{Diffusion Map Plot}
\usage{
PlotDMAP(crs.o, type = c("DPT", "Stage"), stage = NULL, col = NULL)
}
\arguments{
\item{crs.o}{The output of CompCancerRisk.}

\item{type}{The type of diffusion map desired. If DPT, then
cells are colored by diffusion pseudotime. If type
is Stage, then user needs to specify the disease
stage vector with stages labeled by integers. In this
option the user must also specify a color vector to
label each disease stage.}

\item{stage}{An integer vector specifying the stage in cancer progression.}

\item{col}{A color vector specifying the color for each unique disease stage.}
}
\value{
out
A logical.
}
\description{
This function takes the output of CompCancerRisk function and
generates a Diffusion Map plot, indicating root-state and the
cancer-free and cancer-fate tip points.
}
\examples{

}
\references{

}
